// source/frontend/src/models/api/generated/ApiGen_CodeTypes_AddressUsageTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_AddressUsageTypes {
  BILLING = 'BILLING',
  ETLUNKN = 'ETLUNKN',
  MAILING = 'MAILING',
  RESIDNT = 'RESIDNT',
  UNKNOWN = 'UNKNOWN',
}